requirements.txt
torch>=2.0
torchaudio>=2.0
librosa>=0.10
numpy>=1.24
pyyaml>=6.0
onnx>=1.14
onnxscript>=0.6
soundfile>=0.12
sounddevice>=0.4
onnxruntime>=1.16
